FreeBSD: pass address to kmem_free/pmap_unmapdev on 14.0 and higher
This solves a compilation issue on CURRENT where we have a mismatch in
the first argument type for pmap_unmapdev and some kmalloc functions. As
of commit 7ae99f80b6661760c5de3edd330b279f04b092a2 the API changed, so
we should check the FreeBSD version and cast if needed.